fix get_motivation code
fix quotes arrangement on motivational_quotes
