+++ lib/commands/get_workout.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require './lib/workouts/workouts.rb'
+require './lib/data/workouts.rb'
 
 module OfficeWorkout
   module Commands
@@ -8,7 +8,7 @@ module OfficeWorkout
     class GetWorkout < SlackRubyBot::Commands::Base
       command 'get_workout' do |client, data, _match|
         client.say(channel: data.channel, text: 'I am glad you are here! Enjoy the workout!')
-        client.say(channel: data.channel, text: Motivation.motivational_quotes[rand(0...Workouts.motivational_quotes.length)])
+        client.say(channel: data.channel, text: Workouts.workouts_list[rand(0...Workouts.workouts_list.length)])
       end
     end
   end
