+++ office_workout.rb
@@ -2,4 +2,5 @@
 
 require 'slack-ruby-bot'
 require 'lib/commands/get_workout'
+require 'lib/commands/get_motivation'
 require 'lib/bot'
